+++ sdk/core/azure-core-serializer-json-gson/src/main/java/com/azure/core/serializer/json/gson/GsonJsonSerializerProvider.java
@@ -9,7 +9,7 @@ import com.azure.core.util.serializer.MemberNameConverterProvider;
 /**
  * Implementation of {@link JsonSerializerProvider}.
  */
-public class GsonJsonSerializerProvider implements MemberNameConverterProvider, JsonSerializerProvider {
+public final class GsonJsonSerializerProvider implements JsonSerializerProvider, MemberNameConverterProvider {
     @Override
     public GsonJsonSerializer createInstance() {
         return new GsonJsonSerializerBuilder().build();