+++ sdk/resourcemanager/azure-resourcemanager-appservice/src/main/java/com/azure/resourcemanager/appservice/fluent/RecommendationsClient.java
@@ -25,7 +25,6 @@ import com.azure.core.http.rest.RestProxy;
 import com.azure.core.util.Context;
 import com.azure.core.util.FluxUtil;
 import com.azure.core.util.logging.ClientLogger;
-import com.azure.resourcemanager.appservice.WebSiteManagementClient;
 import com.azure.resourcemanager.appservice.fluent.inner.RecommendationCollectionInner;
 import com.azure.resourcemanager.appservice.fluent.inner.RecommendationInner;
 import com.azure.resourcemanager.appservice.fluent.inner.RecommendationRuleInner;