+++ sdk/resourcemanager/azure-resourcemanager-authorization/src/main/java/com/azure/resourcemanager/authorization/fluent/RoleDefinitionsClient.java
@@ -28,7 +28,6 @@ import com.azure.core.management.exception.ManagementException;
 import com.azure.core.util.Context;
 import com.azure.core.util.FluxUtil;
 import com.azure.core.util.logging.ClientLogger;
-import com.azure.resourcemanager.authorization.AuthorizationManagementClient;
 import com.azure.resourcemanager.authorization.fluent.inner.RoleDefinitionInner;
 import com.azure.resourcemanager.authorization.fluent.inner.RoleDefinitionListResultInner;
 import com.azure.resourcemanager.resources.fluentcore.collection.InnerSupportsDelete;