+++ sdk/resourcemanager/azure-resourcemanager-redis/src/main/java/com/azure/resourcemanager/redis/fluent/PatchSchedulesClient.java
@@ -28,7 +28,6 @@ import com.azure.core.management.exception.ManagementException;
 import com.azure.core.util.Context;
 import com.azure.core.util.FluxUtil;
 import com.azure.core.util.logging.ClientLogger;
-import com.azure.resourcemanager.redis.RedisManagementClient;
 import com.azure.resourcemanager.redis.fluent.inner.RedisPatchScheduleInner;
 import com.azure.resourcemanager.redis.fluent.inner.RedisPatchScheduleListResultInner;
 import com.azure.resourcemanager.redis.models.DefaultName;