+++ sdk/resourcemanager/azure-resourcemanager-sql/src/main/java/com/azure/resourcemanager/sql/fluent/WorkloadClassifiersClient.java
@@ -31,7 +31,6 @@ import com.azure.core.util.FluxUtil;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.core.util.polling.PollerFlux;
 import com.azure.core.util.polling.SyncPoller;
-import com.azure.resourcemanager.sql.SqlManagementClient;
 import com.azure.resourcemanager.sql.fluent.inner.WorkloadClassifierInner;
 import com.azure.resourcemanager.sql.fluent.inner.WorkloadClassifierListResultInner;
 import java.nio.ByteBuffer;