+++ sdk/storage/azure-storage-file-datalake/src/test/java/com/azure/storage/file/datalake/FileAPITest.groovy
@@ -2859,7 +2859,6 @@ class FileAPITest extends APISpec {
     }
 
     @Unroll
-    @Requires({ playbackMode() }) // TODO (rickle-msft): Remove annotation
     def "Query min"() {
         setup:
         FileQueryDelimitedSerialization ser = new FileQueryDelimitedSerialization()
@@ -2905,7 +2904,6 @@ class FileAPITest extends APISpec {
     }
 
     @Unroll
-    @Requires({ playbackMode() }) // TODO (rickle-msft): Remove annotation
     def "Query csv serialization separator"() {
         setup:
         FileQueryDelimitedSerialization ser = new FileQueryDelimitedSerialization()
@@ -2981,7 +2979,6 @@ class FileAPITest extends APISpec {
     }
 
     @Unroll
-    @Requires({ playbackMode() }) // TODO (rickle-msft): Remove annotation
     def "Query csv serialization escape and field quote"() {
         setup:
         FileQueryDelimitedSerialization ser = new FileQueryDelimitedSerialization()
@@ -3022,7 +3019,6 @@ class FileAPITest extends APISpec {
 
     /* Note: Input delimited tested everywhere else. */
     @Unroll
-    @Requires({ playbackMode() }) // TODO (rickle-msft): Remove annotation
     def "Query Input json"() {
         setup:
         FileQueryJsonSerialization ser = new FileQueryJsonSerialization()
@@ -3064,7 +3060,6 @@ class FileAPITest extends APISpec {
         1000      | '\n'            || _
     }
 
-    @Requires({ playbackMode() }) // TODO (rickle-msft): Remove annotation
     def "Query Input csv Output json"() {
         setup:
         FileQueryDelimitedSerialization inSer = new FileQueryDelimitedSerialization()
@@ -3105,7 +3100,6 @@ class FileAPITest extends APISpec {
         }
     }
 
-    @Requires({ playbackMode() }) // TODO (rickle-msft): Remove annotation
     def "Query Input json Output csv"() {
         setup:
         FileQueryJsonSerialization inSer = new FileQueryJsonSerialization()
@@ -3146,7 +3140,6 @@ class FileAPITest extends APISpec {
         }
     }
 
-    @Requires({ playbackMode() }) // TODO (rickle-msft): Remove annotation
     def "Query Input csv Output arrow"() {
         setup:
         FileQueryDelimitedSerialization inSer = new FileQueryDelimitedSerialization()
@@ -3183,7 +3176,6 @@ class FileAPITest extends APISpec {
         Base64.getEncoder().encodeToString(osData) == expectedData
     }
 
-    @Requires({ playbackMode() }) // TODO (rickle-msft): Remove annotation
     def "Query non fatal error"() {
         setup:
         FileQueryDelimitedSerialization base = new FileQueryDelimitedSerialization()
@@ -3222,7 +3214,6 @@ class FileAPITest extends APISpec {
         receiver.numErrors > 0
     }
 
-    @Requires({ playbackMode() }) // TODO (rickle-msft): Remove annotation
     def "Query fatal error"() {
         setup:
         FileQueryDelimitedSerialization base = new FileQueryDelimitedSerialization()
@@ -3253,7 +3244,6 @@ class FileAPITest extends APISpec {
         thrown(Exceptions.ReactiveException)
     }
 
-    @Requires({ playbackMode() }) // TODO (rickle-msft): Remove annotation
     def "Query progress receiver"() {
         setup:
         FileQueryDelimitedSerialization base = new FileQueryDelimitedSerialization()
@@ -3297,7 +3287,6 @@ class FileAPITest extends APISpec {
     }
 
     @Requires( { liveMode() } ) // Large amount of data.
-    @Ignore // TODO (rickle-msft): Remove annotation
     def "Query multiple records with progress receiver"() {
         setup:
         FileQueryDelimitedSerialization ser = new FileQueryDelimitedSerialization()
@@ -3349,7 +3338,6 @@ class FileAPITest extends APISpec {
     }
 
     @Unroll
-    @Requires({ playbackMode() }) // TODO (rickle-msft): Remove annotation
     def "Query input output IA"() {
         setup:
         /* Mock random impl of QQ Serialization*/
@@ -3383,7 +3371,6 @@ class FileAPITest extends APISpec {
         false   | true     || _
     }
 
-    @Requires({ playbackMode() }) // TODO (rickle-msft): Remove annotation
     def "Query arrow input IA"() {
         setup:
         def inSer = new FileQueryArrowSerialization()
@@ -3406,8 +3393,24 @@ class FileAPITest extends APISpec {
         thrown(IllegalArgumentException)
     }
 
+    def "Query error"() {
+        setup:
+        fc = fsc.getFileClient(generatePathName())
+
+        when:
+        fc.openQueryInputStream("SELECT * from BlobStorage") /* Don't need to call read. */
+
+        then:
+        thrown(DataLakeStorageException)
+
+        when:
+        fc.query(new ByteArrayOutputStream(), "SELECT * from BlobStorage")
+
+        then:
+        thrown(DataLakeStorageException)
+    }
+
     @Unroll
-    @Requires({ playbackMode() }) // TODO (rickle-msft): Remove annotation
     def "Query AC"() {
         setup:
         match = setupPathMatchCondition(fc, match)
@@ -3449,7 +3452,6 @@ class FileAPITest extends APISpec {
     }
 
     @Unroll
-    @Requires({ playbackMode() }) // TODO (rickle-msft): Remove annotation
     def "Query AC fail"() {
         setup:
         setupPathLeaseCondition(fc, leaseID)