+++ sdk/storage/azure-storage-file-share/CHANGELOG.md
@@ -5,7 +5,7 @@
 - Fixed a bug where snapshot would be appended to a share snapshot instead of sharesnapshot.
 
 - Added support for the 2020-02-10 service version. 
-- Added support to getFileRanges on a previous snapshot. 
+- Added support to getFileRanges on a previous snapshot by adding the getFileRangesDiff API. 
 - Added support to set whether or not smb multichannel is enabled.
 - Added support to lease shares and snapshot shares.
 - Added support to specify a lease id for share operations.